AaharSense AI/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-generativeai==0.8.4
firebase-admin==6.6.0
python-dotenv==1.0.1
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
